# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateSubscriptionRequest:
    properties:
      amount:
        type: number
      auto_renew:
        type: boolean
      billing_cycle:
        enum:
        - weekly
        - monthly
        - quarterly
        - yearly
        - custom
        - one_time
        type: string
      billing_interval:
        type: integer
      category_id:
        type: integer
      currency:
        maxLength: 10
        type: string
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        type: string
      payment_method:
        maxLength: 50
        type: string
      remark:
        maxLength: 500
        type: string
      start_date:
        type: string
      status:
        enum:
        - active
        - paused
        - cancelled
        - expired
        type: string
      website_url:
        maxLength: 500
        type: string
    required:
    - amount
    - billing_cycle
    - name
    type: object
...
  dto.SubscriptionResponse:
    properties:
      amount:
        type: number
      auto_renew:
        type: boolean
      billing_cycle:
        type: string
      billing_interval:
        type: integer
      category_id:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      next_payment_date:
        type: string
      payment_method:
        type: string
      remark:
        type: string
      start_date:
        type: string
      status:
        type: string
      updated_at:
        type: string
      website_url:
        type: string
    type: object
...
  dto.UpdatePasswordRequest:
    properties:
      new_password:
        maxLength: 50
        minLength: 6
        type: string
      old_password:
        maxLength: 50
        minLength: 6
        type: string
    required:
    - new_password
    - old_password
    type: object
  dto.UpdateProfileRequest:
    properties:
      avatar:
        maxLength: 500
        type: string
      nickname:
        maxLength: 50
        type: string
    type: object
  dto.UpdateSubscriptionRequest:
    properties:
      amount:
        type: number
      auto_renew:
        type: boolean
      billing_cycle:
        enum:
        - weekly
        - monthly
        - quarterly
        - yearly
        - custom
        - one_time
        type: string
      billing_interval:
        type: integer
      category_id:
        type: integer
      currency:
        maxLength: 10
        type: string
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        type: string
      payment_method:
        maxLength: 50
        type: string
      remark:
        maxLength: 500
        type: string
      start_date:
        type: string
      status:
        enum:
        - active
        - paused
        - cancelled
        - expired
        type: string
      website_url:
        maxLength: 500
        type: string
    type: object
  dto.UserResponse:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      username:
        type: string
    type: object
  response.PageResult:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Digital asset and subscription management platform API
  title: StackBill API
  version: "1.0"
paths:
...
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Get paginated list of subscriptions with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by category ID
        in: query
        name: category_id
        type: integer
      - description: Filter by status (active, paused, cancelled, expired)
        in: query
        name: status
        type: string
      - description: Filter upcoming renewals only
        in: query
        name: upcoming_renewal
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/dto.SubscriptionResponse'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription
      parameters:
      - description: Subscription data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SubscriptionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Get a single subscription by its ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SubscriptionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update an existing subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subscription data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SubscriptionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update subscription
      tags:
      - subscriptions
  /users/password:
    put:
      consumes:
      - application/json
      description: Change the currently authenticated user's password
      parameters:
      - description: Password data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update password
      tags:
      - user
  /users/profile:
    put:
      consumes:
      - application/json
      description: Update the currently authenticated user's nickname and avatar
      parameters:
      - description: Profile data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"